Update wholeRowElts to long to prevent overflow on large files
Large files <~17Gpx caused an overflow in wholeRowElts and subsequent segfault in std::translate.  Changed to long

+++ pdal/GDALUtils.hpp
@@ -481,8 +481,8 @@ private:
         for (int row = 0; row < yHeight; ++row)
         {
             // Find the offset location in the source container.
-            int wholeRowElts = m_xTotalSize * ((y * m_yBlockSize) + row);
-            int partialRowElts = m_xBlockSize * x;
+            long wholeRowElts = (long)m_xTotalSize * ((y * m_yBlockSize) + row);
+            long partialRowElts = (long)m_xBlockSize * x;
 
             auto si = sourceBegin + (wholeRowElts + partialRowElts);
             std::transform(si, si + xWidth, di,
